391: Cleanup unused code
